Compute degree days

Compute degree days from daily min and max temperature

dd_vals(
  daily_min,
  daily_max,
  nextday_min = NULL,
  thresh_low = NULL,
  thresh_up = NULL,
  method = c("sng_tri", "dbl_tri", "sng_sine", "dbl_sine", "simp_avg")[0],
  cutoff = c("horizontal", "vertical")[1],
  cumulative = FALSE,
  noneg = TRUE,
  debug = FALSE
)

dd_sng_tri(
  daily_min,
  daily_max,
  thresh_low = NULL,
  thresh_up = NULL,
  cumulative = FALSE
)

dd_sng_sine(
  daily_min,
  daily_max,
  thresh_low = NULL,
  thresh_up = NULL,
  cumulative = FALSE
)

dd_dbl_tri(
  daily_min,
  daily_max,
  nextday_min = NULL,
  thresh_low = NULL,
  thresh_up = NULL,
  cumulative = FALSE
)

dd_dbl_sine(
  daily_min,
  daily_max,
  nextday_min = NULL,
  thresh_low = NULL,
  thresh_up = NULL,
  cumulative = FALSE
)

Arguments

daily_min
The daily minimum temperature

daily_max
The daily maximum temperature

nextday_min
The minimum temp the day after

thresh_low
The lower development threshold

thresh_up
The upper development threshold

method
The method used

cutoff
The cutoff method

cumulative
Return cumulative values

noneg
Set negative values to zero

debug
Show additional messages

Details

Units for daily_min, daily_max, thresh_low, and thresh_up should all be the same (i.e., all Fahrenheit or all Celsius). The function does not check for unit consistency.

nextday_min is required for the double-triangle and the double-sine methods. These methods use the minimum temperature of the following day to model temperatures in the 2nd half of thd day.

noneg = TRUE sets negative values to zero. This is generally accepted when using degree days to predict the timing of development milestones, if one assumes that growth can not go backwards.

Functions

dd_sng_tri: Compute degree days using the single-triangle method

dd_sng_sine: Compute degree days using the single-sine method

dd_dbl_tri: Compute degree days using the double-triangle method

dd_dbl_sine: Compute degree days using the double-sine method